Clicking "View query" on Issue categories opens settings.json Fixes #6855
